Workaround substitution bug with notepad.exe
If a substitution was triggered on a line in Notepad, it would cause the
font to revert to Arial for that line. Doesn't happen to a few other
programs I tried, so it seems like a bug specific to Notepad. Caused by
a sub-table having the same name as a lookup table, so they have to have
unique names.